+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/asset_publisher.ftl
@@ -1,5 +1,5 @@
 <#assign
-	assetClassNameIds = dataFactory.assetClassNameIds
+	assetClassNameIds = classNameDataFactory.assetClassNameIds
 	assetVocabularyModels = dataFactory.newAssetVocabularyModels(groupId)
 	pageCounts = dataFactory.getSequence(dataFactory.maxAssetPublisherPageCount)
 />
