+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/ddl.ftl
@@ -1,5 +1,5 @@
 <#assign
-	ddlRecordSetClassNameId = dataFactory.getClassNameId("DDLRecordSet")
+	ddlRecordSetClassNameId = classNameDataFactory.getClassNameId("DDLRecordSet")
 	ddlRecordSetCounts = dataFactory.getSequence(dataFactory.maxDDLRecordSetCount)
 />
