+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_dl_file_type.ftl
@@ -1,6 +1,6 @@
 ${dataFactory.toInsertSQL(dataFactory.newDLFileEntryTypeModel())}
 
-<#assign defaultDLDDMStructureModel = dataFactory.newDefaultDLDDMStructureModel(dataFactory.getClassNameId("DLFileEntry")) />
+<#assign defaultDLDDMStructureModel = dataFactory.newDefaultDLDDMStructureModel(classNameDataFactory.getClassNameId("DLFileEntry")) />
 
 <@insertDDMStructure
 	_ddmStructureLayoutModel=dataFactory.newDefaultDLDDMStructureLayoutModel()
